add folder dict_queue
