+++ implementations/dict_queue/__init__.py
@@ -1,0 +1,5 @@
+from dictq import DictQueue
+from qitem import Item
+
+
+__all__ = ['DictQueue', 'Item']
