+++ pyproject.toml
@@ -13,7 +13,3 @@ dependencies = ["requests", "numpy"]
 
 [project.optional-dependencies]
 dict_queues = []
-another_implementation = [
-    # Dependencies for another implementation
-    "dependency_for_another_implementation",
-]
